refactor(e2e): consolidate duplicated helper functions into common.rs
Move signed_request, signed_request_with_client, signed_request_with_session_token,
and admin_create_user from replication_extension_test.rs and object_lambda_test.rs
into the shared common.rs module.

- signed_request: was duplicated in replication_extension_test.rs and object_lambda_test.rs
- signed_request_with_client: was only in replication_extension_test.rs
- signed_request_with_session_token: was only in replication_extension_test.rs
- admin_create_user: was in replication_extension_test.rs (the existing_object_tag_policy_test.rs
  version stays local as it uses awscurl_put instead of signed_request)

Also cleaned up now-unused imports (sign_v4, UNSIGNED_PAYLOAD, Body, HOST, CONTENT_TYPE)
from the source files.
